requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
cloudinary==1.23.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.3
django-cors-headers==3.5.0
django-environ==0.4.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
drf-yasg==1.20.0
huey==2.3.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
psycopg2-binary==2.8.6
pyparsing==2.4.7
pytz==2020.4
redis==3.5.3
requests==2.25.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
sentry-sdk==0.19.4
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
